+++ packages/uniforms-bootstrap4/src/components/fields/DateField.js
@@ -2,7 +2,7 @@ import React          from 'react';
 import classnames     from 'classnames';
 import {connectField} from 'uniforms';
 
-import FormGroup from './FormGroup';
+import wrapField from '../../lib/wrapField';
 
 const dateFormat = value => value && value.toISOString().slice(0, -8);
 const dateParse = (timestamp, onChange) => {
@@ -13,7 +13,7 @@ const dateParse = (timestamp, onChange) => {
 };
 
 const Date_ = props =>
-    <FormGroup {...props}>
+    wrapField(props, (
         <input
             className={classnames(props.inputClassName, 'form-control', {'form-control-danger': props.error})}
             disabled={props.disabled}
@@ -27,7 +27,7 @@ const Date_ = props =>
             type="datetime-local"
             value={dateFormat(props.value)}
         />
-    </FormGroup>
+    ))
 ;
 
 Date_.displayName = 'Date';